Track child time for each factory
This commit introduces a frame object, and tracks the depth of frames in
on stack so we can record the time spend in any child factories
(associations, callbacks, etc.).

This way we can report both the total time for a factory as well as the
separated self and child times.

The spec we wrote here isn't really a spec (there are no assertions),
and it'd be nice if we could avoid sleeping. But it was helpful in
writing this code so I'll leave it here for inspiration when we write
some proper tests.

Co-authored-by: Adam Hess <hparker@github.com>