Get total by summing self times
Now that we've got access to the self time for each factory (i.e. total
time - time spent building associations), we don't need to hold onto an
extra value for the total time.

Also, the total time was wrong anyway because of an off-by-one error
with depth. The check should have been `depth.zero?`.