Allow injecting a custom collector
This allows customizing both the reporter and collector together (since
the reporter can only call methods defined on the collector). The
collector needs to respond to `collect` with the `frame` as an argument.
We could add quite a bit more info to the frame if we want to.

I'm not sure the name collector is right...