Rename Collector to AggregateStats
The name Collector seems off to me. There's one method for collecting
stats, but the bulk of the class is sorting and filtering the various
FactoryStats for use in the Reporter. John suggested AggregateStats,
which I like a lot more and seems to fit with what the class is doing.