Extract class for child stats
I'd like to be able to report both total and average times for child
stats like we do for top-level stats. The easiest way to make this
happen was to introduce a common base class for Stats that includes the
totals as well as the method for calculating the average.

This also cleans up some code where we were iterating through child
counts and times separately—now all the child data is in one place.

I had to play the same Marshal game as I did with AggregateStats, since
I've got a hash with a default proc.