Introduce block methods to manage subscription
This commit introduces `.reporting` as the primary public method for
using this library, and `.subscribed` when more customized reporting is
needed.

`subscribe` is still available, in case anyone wants a custom stat
object or needs more control over the start and end of the subscription.
The method now returns a `Subscription`, which offers access to the
stats and method for unsubscribing from the notification.