Remove subscribed method
I'm waffling on the api a bit. I wanted `reporting` to return the result
of the block, since I think that's the least surprising behavior, but
the existing `subscribed` didn't make that easy. It could return an array
with stats and the result, but that seemed fairly awkward.

Instead of having a weird api that I want to change later, I'm going to
delete this method. It's still possible to subscribe and report
manually—it'll just be up to the caller to also unsubscribe once they
are done.