Make SimpleReport line count configurable
I've already changed the counts here a couple times, and now other folks
are wanting different counts as well.

This commit makes the count configurable by passing an option to
the report method.

Rather than treating io and count as special configurable properties,
now we forward the whole options hash to the report class. This makes it
possible to pass arbitrary options to any custom reporter classes as
well.